backend/requirements.txt
fastapi==0.115.12
pdfplumber==0.11.6
protobuf==6.30.2
pydantic==2.11.4
python-dotenv==1.1.0
Requests==2.32.3
